Developer reference

Developer reference

SDPAFamily.presolve — Method.

presolve(optimizer::SDPAFamily.Optimizer{T}) where T

Identifies linearly dependent constraints in the problem. This is done by a naive Gaussian elimination.

Returns a vector with the indices of redundant constraints, which should be removed from the formulation. The rest of the constraints form a maximal linearly independent subset of the original set of constraints.

source

SDPAFamily.read_results! — Method.

read_results!(optimizer::Optimizer{T}, filepath::String, redundant_entries::Vector)

Populates optimizer with results in a SDPA-formatted output file specified by filepath. Redundant entries corresponding to linearly dependent constraints are set to 0.

source

SDPAFamily.BB_PATHS — Constant.

BB_PATHS::Dict{Symbol,String}

Holds the binary-builder-built paths to the executables for sdpa_gmp, sdpa_dd, and sdpa_qd.

source

SDPAFamily.WSLize_path — Method.

WSLize_path(path::String) -> String

This function converts Windows paths for use via WSL.

source

SDPAFamily.initializeSolve — Method.

initializeSolve(optimizer::Optimizer)

Writes problem data into an SDPA-formatted file named input.dat-s. presolve.jl routine is applied as indicated by optimizer.presolve.

Returns a vector of indices for redundant constraints, which are omitted from the input file.

source

SDPAFamily.inputElement — Method.

inputElement(optimizer::Optimizer, constr_number::Int, blk::Int, i::Int, j::Int, value::T) where T

Stores the constraint data in optimizer.elemdata as a vector of tuples. Each tuple corresponds to one line in the SDPA-formatted input file.

source

SDPAFamily.reduce! — Method.

function reduce!(A::SparseMatrixCSC{T}, ɛ = T <: Union{Rational,Integer} ? 0 : eps(norm(A, Inf))) where T

Identifies linearly dependent constraints in the problem. The last column of input is constraint constants and they are included to check if the linearly dependent constraints are redundant or inconsistent. This is done by a naive Gaussian elimination.

Returns a vector with the indices of redundant constraints, which should be removed from the formulation. The rest of the constraints form a maximal linearly independent subset of the original set of constraints.

source

SDPAFamily.sdpa_gmp_binary_solve! — Method.

sdpa_gmp_binary_solve!(m::Optimizer, full_input_path, full_output_path; extra_args::Cmd, redundant_entries)

Calls the binary sdpa_gmp to solve SDPA-formatted problem specified in a .dat-s file at full_input_path. Results are written into the file at full_output_path. extra_args is passed on to the binary as additional options, allowing for e.g. custom parameter files. redundant_entries is a sorted vector listing indices of linearly dependent constraint which are already removed by presolve.jl. The corresponding decision variables are populated as zeros.

This function returns m with solutions already populated from results in the output file.

source